#3 Call Activity -> Fragment 로 변경 및 디자인 수정
