+++ app/src/main/java/kr/hs/emirim/s2019w28/dongji/CallActivity.java
@@ -1,45 +1,45 @@
-package kr.hs.emirim.s2019w28.dongji;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.media.Image;
-import android.net.Uri;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.ImageButton;
-
-public class CallActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_call);
-
-        ImageButton btn_1339 = findViewById(R.id.Bogun);
-        btn_1339.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent();
-                intent.setAction(Intent.ACTION_DIAL);
-                intent.setData(Uri.parse("tel:1339"));
-                startActivity(intent);
-            }
-        });
-
-        ImageButton btn_054 = findViewById(R.id.Nong);
-        btn_054.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent();
-                intent.setAction(Intent.ACTION_DIAL);
-                intent.setData(Uri.parse("tel:054-912-0616"));
-                startActivity(intent);
-            }
-        });
-    }
-
-
-
-
-}
+//package kr.hs.emirim.s2019w28.dongji;
+//
+//import androidx.appcompat.app.AppCompatActivity;
+//
+//import android.content.Intent;
+//import android.media.Image;
+//import android.net.Uri;
+//import android.os.Bundle;
+//import android.view.View;
+//import android.widget.ImageButton;
+//
+//public class CallActivity extends AppCompatActivity {
+//
+//    @Override
+//    protected void onCreate(Bundle savedInstanceState) {
+//        super.onCreate(savedInstanceState);
+//        setContentView(R.layout.activity_call);
+//
+//        ImageButton btn_1339 = findViewById(R.id.Bogun);
+//        btn_1339.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                Intent intent = new Intent();
+//                intent.setAction(Intent.ACTION_DIAL);
+//                intent.setData(Uri.parse("tel:1339"));
+//                startActivity(intent);
+//            }
+//        });
+//
+//        ImageButton btn_054 = findViewById(R.id.Nong);
+//        btn_054.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                Intent intent = new Intent();
+//                intent.setAction(Intent.ACTION_DIAL);
+//                intent.setData(Uri.parse("tel:054-912-0616"));
+//                startActivity(intent);
+//            }
+//        });
+//    }
+//
+//
+//
+//
+//}
