+++ app/src/main/java/kr/hs/emirim/s2019w28/dongji/fragment/CallFragment.java
@@ -1,5 +1,7 @@
 package kr.hs.emirim.s2019w28.dongji.fragment;
 
+import android.content.Intent;
+import android.net.Uri;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -7,60 +9,56 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageButton;
+import android.widget.TextView;
 
 import kr.hs.emirim.s2019w28.dongji.R;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link CallFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
+
 public class CallFragment extends Fragment {
-
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-
-    public CallFragment() {
-        // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment Call.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static CallFragment newInstance(String param1, String param2) {
-        CallFragment fragment = new CallFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
 
+    }
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_call, container, false);
+        View view = inflater.inflate(R.layout.fragment_call, null);
+
+        ImageButton btn_1339 = view.findViewById(R.id.Bogun);
+        btn_1339.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent();
+                intent.setAction(Intent.ACTION_DIAL);
+                intent.setData(Uri.parse("tel:1339"));
+                startActivity(intent);
+            }
+        });
+
+        ImageButton btn_054 = view.findViewById(R.id.Nong);
+        btn_054.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent();
+                intent.setAction(Intent.ACTION_DIAL);
+                intent.setData(Uri.parse("tel:054-912-0616"));
+                startActivity(intent);
+            }
+        });
+
+        ImageButton btn_1588 = view.findViewById(R.id.Ga);
+        btn_1588.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent();
+                intent.setAction(Intent.ACTION_DIAL);
+                intent.setData(Uri.parse("tel:1588-9060"));
+                startActivity(intent);
+            }
+        });
+        return view;
     }
 }
